requirements.txt
Django==2.1.7
django-choices==1.6.2
django-dynamic-fixture==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
pytz==2018.9
six==1.12.0
